chore: more fixes and refactors
