+++ src/components/tables/BulkEditTable/BulkEditTable.module.less
@@ -6,6 +6,7 @@
   td {
     max-width: 150px;
     text-overflow: ellipsis;
+    overflow-wrap: break-word;
     vertical-align: middle;
 
     &.dataColumn {
