+++ src/i18n/en.json
@@ -1155,18 +1155,18 @@
     "add_event": "Add Activity",
     "title": "Title",
     "type": "Activity Type",
-    "event_type_capacity": "Select..",
-    "selection_plan": "Select Selection Plan",
-    "location": "Select Location",
-    "level": "Select Level",
+    "event_type_capacity": "Activity Type Capacity",
+    "selection_plan": "Selection Plan",
+    "location": "Location",
+    "level": "Activity Level",
     "tags": "Tags",
     "progress_flags": "Progress Flags",
     "streaming_url": "Streaming URL",
     "meeting_url": "Meeting URL",
     "etherpad_link": "Etherpad URL",
     "streaming_type": "Streaming Type",
-    "selection_status": "Select Status",
-    "submission_status": "Select Status",
+    "selection_status": "Selection Status",
+    "submission_status": "Submission Status",
     "published": "Published",
     "speakers": "Speakers",
     "duration": "Duration",
@@ -1175,12 +1175,14 @@
     "submitter_company": "Submitter Company",
     "sponsor": "Sponsors",
     "all_companies": "All Companies",
-    "track": "Select Activity Category",
+    "track": "Activity Category",
     "start_date": "Start Date",
     "end_date": "End Date",
     "media_uploads_display": "Display on Site",
     "allow_feedback": "Allow feedback ?",
     "to_record": "To record ?",
+    "created_by": "Submitter",
+    "review_status": "Review Status",
     "delete_event_warning": "Are you sure you want to delete activity",
     "import": "Import",
     "import_events": "Import Activity",
@@ -1193,14 +1195,12 @@
     "mux_token_secret_info": "MUX Token Secret",
     "mux_email_to": "Email to",
     "mux_email_to_info": "Recipient to send process excerpt.",
-    "created_by": "Submitter",
     "mux_import_done": "MUX MP4 Import Process Requested Successfuly.",
     "allows_attendee_vote_filter": "Allows attendee vote",
     "allows_location_filter": "Allows location",
     "allows_publishing_dates_filter": "Allows publishing dates",
     "apply_filters": "Apply Filters",
     "select_fields": "Show Columns",
-    "review_status": "Review Status",
     "created": "Created",
     "modified": "Modified",
     "submission_source": "Submission Source",
