+++ src/pages/events/summit-event-list-page/index.js
@@ -15,21 +15,13 @@ import React, { useEffect, useState } from "react";
 import { connect } from "react-redux";
 import T from "i18n-react/dist/i18n-react";
 import Swal from "sweetalert2";
-import moment from "moment-timezone";
 import { Pagination } from "react-bootstrap";
 import Dropdown from "openstack-uicore-foundation/lib/components/inputs/dropdown";
 import FreeTextSearch from "openstack-uicore-foundation/lib/components/free-text-search";
-import SpeakerInput from "openstack-uicore-foundation/lib/components/inputs/speaker-input";
-import { escapeFilterValue } from "openstack-uicore-foundation/lib/utils/actions";
 import {
   GridFilter,
-  OPERATORS,
   useGridFilter
 } from "openstack-uicore-foundation/lib/components/mui/grid-filter";
-import {
-  queryMembers,
-  queryTags
-} from "openstack-uicore-foundation/lib/utils/query-actions";
 import {
   bulkUpdateEvents,
   changeEventListSearchTerm,
@@ -37,10 +29,7 @@ import {
   exportEvents,
   getEvents,
   importEventsCSV,
-  importMP4AssetsFromMUX,
-  queryAllCompanies,
-  querySpeakerCompany,
-  querySubmitterCompany
+  importMP4AssetsFromMUX
 } from "../../../actions/event-actions";
 import { getMediaUploads } from "../../../actions/media-upload-actions";
 import { handleDDLSortByLabel } from "../../../utils/methods";
@@ -48,10 +37,8 @@ import {
   DEFAULT_CURRENT_PAGE,
   DEFAULT_Z_INDEX,
   HIGH_Z_INDEX,
-  MAX_PER_PAGE,
-  MILLISECONDS_IN_SECOND
+  MAX_PER_PAGE
 } from "../../../utils/constants";
-import { defaultColumns, editableColumns } from "../../../utils/summitUtils";
 import SaveFilterCriteria from "../../../components/filters/save-filter-criteria";
 import SelectFilterCriteria from "../../../components/filters/select-filter-criteria";
 import {
@@ -63,1024 +50,15 @@ import BulkEditTable from "../../../components/tables/BulkEditTable";
 import { buildNameIdDDL } from "../../../utils/events/summit-event-list-page.utils";
 import ImportModal from "./components/ImportModal";
 import ImportMUXModal from "./components/ImportMUXModal";
+import {
+  formatEventData,
+  getCriterias,
+  getOptionalColumns,
+  toApiSortKey,
+  toUiSortKey
+} from "./helpers";
 
 const FILTER_ID = "summit_activity_list";
-
-const formatDuration = (duration) => {
-  const d = moment.duration(duration, "seconds");
-  return d.format("mm:ss") !== "00" ? d.format("mm:ss") : "TBD";
-};
-
-const formatTimestamp = (timestamp, tz) =>
-  timestamp
-    ? moment(timestamp * MILLISECONDS_IN_SECOND)
-        .tz(tz)
-        .format("MMMM Do YYYY, h:mm a")
-    : null;
-
-// Adds display-only fields for the table to render; never overwrites a real
-// event field, since the same row object is sent back as-is to bulkUpdateEvents.
-export const formatEventData = (e, summit) => {
-  let speakers_companies =
-    Array.isArray(e.speakers) && e.speakers.length > 0
-      ? e.speakers.map((s) => s.company)
-      : [];
-  speakers_companies =
-    speakers_companies.length > 0
-      ? speakers_companies.filter(
-          (item, index) =>
-            item !== "" && speakers_companies.indexOf(item) === index
-        )
-      : [];
-
-  const event_type_capacity = [];
-
-  if (e.type?.allows_location) event_type_capacity.push("Allows Location");
-  if (e.type?.allows_attendee_vote)
-    event_type_capacity.push("Allows Attendee Vote");
-  if (e.type?.allows_publishing_dates)
-    event_type_capacity.push("Allows Publishing Dates");
-
-  let speakers_count;
-
-  if (e.type?.use_speakers) {
-    if (e.speakers && e.speakers.length > 0) {
-      speakers_count = e.speakers.length;
-    } else {
-      speakers_count = "0";
-    }
-  } else {
-    speakers_count = "N/A";
-  }
-
-  return {
-    ...e,
-    created_by_fullname: e.hasOwnProperty("created_by")
-      ? `${e.created_by.first_name} ${e.created_by.last_name} (${e.created_by.email})`
-      : "TBD",
-    submitter_company: e.hasOwnProperty("created_by")
-      ? e.created_by.company
-      : "N/A",
-    speaker_names:
-      Array.isArray(e.speakers) && e.speakers.length > 0
-        ? e.speakers.map((s) => `${s.first_name} ${s.last_name}`).join(", ")
-        : "N/A",
-    speaker_company:
-      speakers_companies.length > 0
-        ? speakers_companies.reduce(
-            (accumulator, company) =>
-              accumulator + (accumulator !== "" ? ", " : "") + company,
-            ""
-          )
-        : "N/A",
-    speakers_count,
-    event_type_capacity: event_type_capacity.reduce(
-      (accumulator, capacity) =>
-        accumulator + (accumulator !== "" ? ", " : "") + capacity,
-      ""
-    ),
-    track_name: e?.track?.name ? e.track.name : "TBD",
-    sponsor:
-      Array.isArray(e.sponsors) && e.sponsors.length > 0
-        ? e.sponsors.map((s) => s.name).join(", ")
-        : "N/A",
-    progress_flags: e?.actions
-      ?.map((a) => `${a.type.label} (${a.is_completed ? "ON" : "OFF"})`)
-      .join(", "),
-    published_date_display: e.is_published
-      ? moment(e.published_date * MILLISECONDS_IN_SECOND)
-          .tz(summit.time_zone.name)
-          .format("MMMM Do YYYY, h:mm a")
-      : "No",
-    start_date_display:
-      formatTimestamp(e.start_date, summit.time_zone.name) ?? "TBD",
-    end_date_display:
-      formatTimestamp(e.end_date, summit.time_zone.name) ?? "TBD",
-    created_display: formatTimestamp(e.created, summit.time_zone_id) ?? "TBD",
-    modified: formatTimestamp(e.last_edited, summit.time_zone_id) ?? "TBD"
-  };
-};
-
-const fieldNames = (
-  allSelectionPlans,
-  allTracks,
-  eventTypes,
-  currentSummitId
-) => [
-  {
-    columnKey: "speaker_names",
-    value: "speakers",
-    customStyle: { minWidth: "350px" },
-    editableField: (extraProps) => {
-      const useSpeakers = extraProps.row.type?.use_speakers;
-      return (
-        useSpeakers && (
-          <SpeakerInput
-            id="speakers"
-            isClearable
-            isMulti
-            placeholder={T.translate("edit_event.search_speakers")}
-            menuPortalTarget={document.body}
-            menuPosition="fixed"
-            styles={{
-              menuPortal: (base) => ({ ...base, zIndex: 9999 }),
-              control: (base, state) => ({
-                ...base,
-                zIndex: state.menuIsOpen ? HIGH_Z_INDEX : DEFAULT_Z_INDEX
-              })
-            }}
-            getOptionLabel={(speaker) =>
-              `${speaker.first_name} ${speaker.last_name} (${speaker.email})`
-            }
-            // eslint-disable-next-line react/jsx-props-no-spreading
-            {...extraProps}
-            value={extraProps.row.speakers}
-          />
-        )
-      );
-    }
-  },
-  { columnKey: "created_by_fullname", value: "created_by", sortable: true },
-  {
-    columnKey: "published_date",
-    value: "published",
-    sortable: true,
-    render: (_, row) => row.published_date_display
-  },
-  {
-    columnKey: "duration",
-    value: "duration",
-    sortable: true,
-    render: (duration, row) =>
-      row.type?.allows_publishing_dates && duration
-        ? formatDuration(duration)
-        : "N/A"
-  },
-  { columnKey: "speakers_count", value: "speakers_count", sortable: true },
-  { columnKey: "speaker_company", value: "speaker_company", sortable: true },
-  {
-    columnKey: "track_name",
-    value: "track",
-    sortable: true,
-    editableField: (extraProps) => {
-      const trackOptions = buildNameIdDDL(allTracks);
-
-      return (
-        <Dropdown
-          id="track"
-          options={trackOptions}
-          menuPortalTarget={document.body}
-          menuPosition="fixed"
-          styles={{
-            menuPortal: (base) => ({ ...base, zIndex: 9999 }),
-            control: (base, state) => ({
-              ...base,
-              zIndex: state.menuIsOpen ? HIGH_Z_INDEX : DEFAULT_Z_INDEX
-            })
-          }}
-          // eslint-disable-next-line react/jsx-props-no-spreading
-          {...extraProps}
-          value={extraProps.row.track?.id}
-        />
-      );
-    }
-  },
-  {
-    columnKey: "start_date",
-    value: "start_date",
-    sortable: true,
-    render: (_, row) => row.start_date_display
-  },
-  {
-    columnKey: "end_date",
-    value: "end_date",
-    sortable: true,
-    render: (_, row) => row.end_date_display
-  },
-  { columnKey: "submitters", value: "submitters" },
-  {
-    columnKey: "submitter_company",
-    value: "submitter_company",
-    sortable: true
-  },
-  { columnKey: "sponsor", value: "sponsor", sortable: true },
-  { columnKey: "event_type_capacity", value: "event_type_capacity" },
-  {
-    columnKey: "selection_plan",
-    value: "selection_plan",
-    sortable: true,
-    editableField: (extraProps) => {
-      if (!extraProps.row?.type?.id) return false;
-      const eventType = Array.isArray(eventTypes)
-        ? eventTypes.find(
-            (t) => t?.id !== undefined && t.id === extraProps.row.type?.id
-          )
-        : null;
-      if (!eventType) return false;
-
-      const allowSelectionPlanEdit =
-        ["PresentationType"].includes(eventType.class_name) ||
-        ["PresentationType"].includes(eventType.name);
-      if (!allowSelectionPlanEdit) return false;
-
-      const trackId = extraProps.row?.track?.id;
-      const track =
-        trackId !== undefined && trackId !== null
-          ? allTracks.find((t) => t?.id !== undefined && t.id === trackId)
-          : null;
-
-      const selectionPlansPerTrack = buildNameIdDDL(
-        (Array.isArray(allSelectionPlans) ? allSelectionPlans : []).filter(
-          (sp) =>
-            !track ||
-            (Array.isArray(sp.track_groups) &&
-              Array.isArray(track.track_groups) &&
-              sp.track_groups.some((gr) => track.track_groups.includes(gr)))
-        )
-      );
-
-      return (
-        <Dropdown
-          id="selection_plan"
-          options={selectionPlansPerTrack}
-          value={extraProps.value || ""}
-          menuPortalTarget={document.body}
-          menuPosition="fixed"
-          styles={{
-            menuPortal: (base) => ({ ...base, zIndex: 9999 }),
-            control: (base, state) => ({
-              ...base,
-              width: 220,
-              zIndex: state.menuIsOpen ? HIGH_Z_INDEX : DEFAULT_Z_INDEX
-            })
-          }}
-          // eslint-disable-next-line react/jsx-props-no-spreading
-          {...extraProps}
-        />
-      );
-    },
-    render: (e) => (e?.name ? e.name : "N/A")
-  },
-  {
-    columnKey: "location",
-    value: "location",
-    sortable: true,
-    render: (location) => (location?.name ? location.name : "N/A")
-  },
-  {
-    columnKey: "level",
-    value: "level",
-    sortable: true,
-    render: (level) => level || "N/A"
-  },
-  {
-    columnKey: "tags",
-    value: "tags",
-    sortable: true,
-    render: (tags) =>
-      Array.isArray(tags) && tags.length > 0
-        ? tags.reduce(
-            (accumulator, t) =>
-              accumulator + (accumulator !== "" ? ", " : "") + t.tag,
-            ""
-          )
-        : "N/A"
-  },
-  {
-    columnKey: "streaming_url",
-    value: "streaming_url",
-    sortable: true,
-    title: true,
-    editableField: true,
-    placeholder: T.translate("bulk_actions_page.placeholders.streaming_url"),
-    render: (url) => url || "N/A"
-  },
-  {
-    columnKey: "meeting_url",
-    value: "meeting_url",
-    sortable: true,
-    title: true,
-    editableField: true,
-    placeholder: T.translate("bulk_actions_page.placeholders.meeting_url"),
-    render: (url) => url || "N/A"
-  },
-  {
-    columnKey: "etherpad_link",
-    value: "etherpad_link",
-    sortable: true,
-    title: true,
-    editableField: true,
-    placeholder: T.translate("bulk_actions_page.placeholders.etherpad_link"),
-    render: (link) => link || "N/A"
-  },
-  {
-    columnKey: "streaming_type",
-    value: "streaming_type",
-    sortable: true,
-    render: (type) => type || "N/A"
-  },
-  {
-    columnKey: "review_status",
-    value: "review_status",
-    sortable: true,
-    title: true,
-    render: (status) => status ?? "N/A"
-  },
-  {
-    columnKey: "status",
-    value: "submission_status",
-    sortable: true,
-    title: true,
-    render: (status) => status ?? "Not Submitted"
-  },
-  {
-    columnKey: "progress_flags",
-    value: "progress_flags",
-    sortable: true,
-    title: true
-  },
-  {
-    columnKey: "created",
-    value: "created",
-    sortable: true,
-    render: (_, row) => row.created_display
-  },
-  { columnKey: "modified", value: "modified", sortable: true },
-  {
-    columnKey: "submission_source",
-    value: "submission_source",
-    sortable: true,
-    render: (source) => source || "N/A"
-  },
-  {
-    columnKey: "media_uploads",
-    value: "media_uploads",
-    sortable: false,
-    render: (e, row) => {
-      if (!e?.length) return "N/A";
-      return (
-        <>
-          {e.map((m) => (
-            <React.Fragment key={m.id}>
-              <button
-                type="button"
-                className="text-link-button"
-                onClick={(ev) => {
-                  ev.preventDefault();
-                  if (!row?.id || !currentSummitId) return false;
-                  window.open(
-                    `/app/summits/${currentSummitId}/events/${row.id}/materials/${m.id}`,
-                    "_blank"
-                  );
-                  return false;
-                }}
-              >
-                {m.media_upload_type.name} - {m.created}
-              </button>
-              <br />
-            </React.Fragment>
-          ))}
-        </>
-      );
-    }
-  },
-  {
-    columnKey: "media_uploads_display",
-    value: "media_uploads_display",
-    sortable: false,
-    render: (e, row) => {
-      const mediaUploads = row?.media_uploads || [];
-      if (!mediaUploads.length) return "N/A";
-      return (
-        <>
-          {mediaUploads.map((m) => (
-            <React.Fragment key={m.id}>
-              {`"${m.media_upload_type.name}" : `}
-              <b>{`${m.display_on_site ? "Yes" : "No"}`}</b>
-              <br />
-            </React.Fragment>
-          ))}
-        </>
-      );
-    }
-  },
-  {
-    columnKey: "allow_feedback",
-    value: "allow_feedback",
-    sortable: false,
-    render: (field) =>
-      field === undefined ? "N/A" : field === true ? "Yes" : "No",
-    editableField: (extraProps) => (
-      <Dropdown
-        id="allow_feedback"
-        value={extraProps}
-        options={[
-          { label: "Yes", value: true },
-          { label: "No", value: false }
-        ]}
-        menuPortalTarget={document.body}
-        menuPosition="fixed"
-        styles={{
-          menuPortal: (base) => ({ ...base, zIndex: 9999 }),
-          control: (base, state) => ({
-            ...base,
-            zIndex: state.menuIsOpen ? HIGH_Z_INDEX : DEFAULT_Z_INDEX
-          })
-        }}
-        // eslint-disable-next-line react/jsx-props-no-spreading
-        {...extraProps}
-      />
-    )
-  },
-  {
-    columnKey: "to_record",
-    value: "to_record",
-    sortable: false,
-    render: (field) =>
-      field === undefined ? "N/A" : field === true ? "Yes" : "No",
-    editableField: (extraProps) => (
-      <Dropdown
-        id="to_record"
-        value={extraProps}
-        options={[
-          { label: "Yes", value: true },
-          { label: "No", value: false }
-        ]}
-        menuPortalTarget={document.body}
-        menuPosition="fixed"
-        styles={{
-          menuPortal: (base) => ({ ...base, zIndex: 9999 }),
-          control: (base, state) => ({
-            ...base,
-            zIndex: state.menuIsOpen ? HIGH_Z_INDEX : DEFAULT_Z_INDEX
-          })
-        }}
-        // eslint-disable-next-line react/jsx-props-no-spreading
-        {...extraProps}
-      />
-    )
-  }
-];
-
-const eventTypeCapacityOptions = [
-  {
-    value: "allows_attendee_vote_filter",
-    label: T.translate("event_list.allows_attendee_vote_filter")
-  },
-  {
-    value: "allows_location_filter",
-    label: T.translate("event_list.allows_location_filter")
-  },
-  {
-    value: "allows_publishing_dates_filter",
-    label: T.translate("event_list.allows_publishing_dates_filter")
-  }
-];
-
-const levelOptions = [
-  { label: "Beginner", value: "beginner" },
-  { label: "Intermediate", value: "intermediate" },
-  { label: "Advanced", value: "advanced" },
-  { label: "N/A", value: "na" }
-];
-
-const selectionStatusOptions = [
-  { label: "Pending", value: "pending" },
-  { label: "Accepted", value: "accepted" },
-  { label: "Rejected", value: "rejected" },
-  { label: "Alternate", value: "alternate" }
-];
-
-const publishedStatusOptions = [
-  {
-    label: "Published",
-    value: "1"
-  },
-  {
-    label: "Non Published",
-    value: "0"
-  }
-];
-
-const rsvpOptions = [
-  {
-    label: "Has RSVP",
-    value: true
-  },
-  {
-    label: "No RSVP",
-    value: false
-  }
-];
-
-const streamingTypeOptions = [
-  { label: "LIVE", value: "LIVE" },
-  { label: "VOD", value: "VOD" }
-];
-
-const submissionStatusOptions = [
-  {
-    label: T.translate("event_list.submission_status_accepted"),
-    value: "Accepted"
-  },
-  {
-    label: T.translate("event_list.submission_status_received"),
-    value: "Received"
-  },
-  {
-    label: T.translate("event_list.submission_status_not_submitted"),
-    value: "NonReceived"
-  }
-];
-
-const reviewStatusOptions = [
-  {
-    label: T.translate("event_list.review_status_accepted"),
-    value: "Accepted"
-  },
-  {
-    label: T.translate("event_list.review_status_in_review"),
-    value: "InReview"
-  },
-  {
-    label: T.translate("event_list.review_status_no_submitted"),
-    value: "NotSubmitted"
-  },
-  {
-    label: T.translate("event_list.review_status_published"),
-    value: "Published"
-  },
-  {
-    label: T.translate("event_list.review_status_received"),
-    value: "Received"
-  },
-  {
-    label: T.translate("event_list.review_status_rejected"),
-    value: "Rejected"
-  }
-];
-
-const submissionSourceOptions = [
-  { label: "Admin", value: "Admin" },
-  { label: "Submission", value: "Submission" }
-];
-
-// queryTags' response shape differs depending on whether it's summit-scoped
-// (nested `tag.tag`) or global (flat `tag`); handle both defensively.
-const tagFormatOption = (item) => ({
-  value: item.id,
-  label:
-    typeof item.tag === "string" ? item.tag : item.tag?.tag || String(item.id)
-});
-
-const getCriterias = (summit, mediaUploadTypes) => [
-  {
-    key: "event_type_capacity",
-    label: "Activity Type Capacity",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: eventTypeCapacityOptions,
-        multiple: true
-      }
-    },
-    customParser: (f) => {
-      const filter = [];
-
-      if (f.value.includes("allows_attendee_vote_filter")) {
-        filter.push("type_allows_attendee_vote==1");
-      }
-      if (f.value.includes("allows_location_filter")) {
-        filter.push("type_allows_location==1");
-      }
-      if (f.value.includes("allows_publishing_dates_filter")) {
-        filter.push("type_allows_publishing_dates==1");
-      }
-
-      return filter;
-    }
-  },
-  {
-    key: "selection_plan_id",
-    label: "Selection Plan",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: summit.selection_plans.map((sp) => ({
-          label: sp.name,
-          value: sp.id
-        })),
-        multiple: true
-      }
-    }
-  },
-  {
-    key: "location_id",
-    label: "Location",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: summit.locations.map((sp) => ({
-          label: sp.name,
-          value: sp.id
-        })),
-        multiple: true
-      }
-    }
-  },
-  {
-    key: "selection_status",
-    label: "Selection Status",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: selectionStatusOptions,
-        multiple: true
-      }
-    }
-  },
-  {
-    key: "track_id",
-    label: "Activity Category",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: summit.tracks.map((t) => ({ label: t.name, value: t.id })),
-        multiple: true
-      }
-    }
-  },
-  {
-    key: "event_type_id",
-    label: "Activity Type",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: summit.event_types.map((type) => ({
-          label: type.name,
-          value: type.id
-        })),
-        multiple: true
-      }
-    }
-  },
-  {
-    key: "speaker_id",
-    label: "Speakers",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "speaker",
-      props: {
-        summitId: summit.id,
-        multiple: true
-      }
-    },
-    customParser: (f) => [
-      `speaker_id==${f.value.map((s) => s.value).join("||")}`
-    ]
-  },
-  {
-    key: "speaker_company",
-    label: "Speaker Company",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "company",
-      props: {
-        queryFunction: querySpeakerCompany,
-        multiple: true
-      }
-    },
-    customParser: (f) => [
-      `speaker_company==${f.value
-        .map((c) => escapeFilterValue(c.raw.name))
-        .join("||")}`
-    ]
-  },
-  {
-    key: "level",
-    label: "Activity Level",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: levelOptions,
-        multiple: true
-      }
-    }
-  },
-  {
-    key: "tags",
-    label: "Tags",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "asyncSelect",
-      props: {
-        queryFunction: (input, callback) =>
-          queryTags(summit.id, input, callback),
-        formatOption: tagFormatOption,
-        multiple: true
-      }
-    },
-    customParser: (f) => [
-      `tags==${f.value.map((t) => escapeFilterValue(t.label)).join("||")}`
-    ]
-  },
-  {
-    key: "published",
-    label: "Published",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: publishedStatusOptions
-      }
-    }
-  },
-  {
-    key: "rsvp_type",
-    label: "Has RSVP?",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: rsvpOptions
-      }
-    },
-    customParser: (f) => [`rsvp_type${f.value ? "<>" : "=="}None`]
-  },
-  {
-    key: "progress_flag",
-    label: "Progress Flag",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: summit.presentation_action_types.map((pf) => ({
-          value: pf.id,
-          label: pf.label
-        })),
-        multiple: true
-      }
-    },
-    customParser: (f) => [
-      f.value.map((pf) => `actions==type_id==${pf}&&is_completed==1`).join(",")
-    ]
-  },
-  {
-    key: "created",
-    label: "Created",
-    operators: [OPERATORS.BEFORE, OPERATORS.AFTER],
-    values: {
-      type: "datetime",
-      props: {
-        mode: "datetime"
-      }
-    }
-  },
-  {
-    key: "last_edited",
-    label: "Modified",
-    operators: [OPERATORS.BEFORE, OPERATORS.AFTER],
-    values: {
-      type: "datetime",
-      props: {
-        mode: "datetime"
-      }
-    }
-  },
-  {
-    key: "start_date",
-    label: "Start Date",
-    operators: [OPERATORS.BEFORE, OPERATORS.AFTER],
-    values: {
-      type: "datetime",
-      props: {
-        mode: "datetime"
-      }
-    }
-  },
-  {
-    key: "end_date",
-    label: "End Date",
-    operators: [OPERATORS.BEFORE, OPERATORS.AFTER],
-    values: {
-      type: "datetime",
-      props: {
-        mode: "datetime"
-      }
-    }
-  },
-  {
-    key: "duration",
-    label: "Duration",
-    operators: [OPERATORS.IS, OPERATORS.LESS, OPERATORS.GREATER],
-    values: {
-      type: "number",
-      props: {
-        min: 0,
-        integer: true
-      }
-    }
-  },
-  {
-    key: "speakers_count",
-    label: "Speaker Count",
-    operators: [OPERATORS.IS, OPERATORS.LESS, OPERATORS.GREATER],
-    values: {
-      type: "number",
-      props: {
-        min: 0,
-        integer: true
-      }
-    }
-  },
-  {
-    key: "submitters",
-    label: "Submitters",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "asyncSelect",
-      props: {
-        queryFunction: queryMembers,
-        formatOption: (m) => ({
-          value: m.id,
-          label: `${m.first_name} ${m.last_name} (${m.email})`
-        }),
-        multiple: true
-      }
-    },
-    customParser: (f) => [
-      f.value
-        .flatMap((s) => {
-          const escapedFullName = escapeFilterValue(
-            `${s.raw.first_name} ${s.raw.last_name}`
-          );
-          const escapedEmail = escapeFilterValue(s.raw.email);
-          const fullNameFilter = `created_by_fullname==${escapedFullName}`;
-          const emailFilter = `created_by_email==${escapedEmail}`;
-          return [fullNameFilter, emailFilter];
-        })
-        .join(",")
-    ]
-  },
-  {
-    key: "created_by_company",
-    label: "Submitter Company",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "company",
-      props: {
-        queryFunction: querySubmitterCompany,
-        multiple: true
-      }
-    },
-    customParser: (f) => [
-      `created_by_company==${f.value
-        .map((c) => escapeFilterValue(c.raw.name))
-        .join("||")}`
-    ]
-  },
-  {
-    key: "streaming_url",
-    label: "Streaming URL",
-    operators: [OPERATORS.IS, OPERATORS.LIKE_START, OPERATORS.LIKE],
-    values: {
-      type: "text",
-      props: {}
-    }
-  },
-  {
-    key: "meeting_url",
-    label: "Meeting URL",
-    operators: [OPERATORS.IS, OPERATORS.LIKE_START, OPERATORS.LIKE],
-    values: {
-      type: "text",
-      props: {}
-    }
-  },
-  {
-    key: "etherpad_link",
-    label: "Etherpad Link",
-    operators: [OPERATORS.IS, OPERATORS.LIKE_START, OPERATORS.LIKE],
-    values: {
-      type: "text",
-      props: {}
-    }
-  },
-  {
-    key: "streaming_type",
-    label: "Streaming Type",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: streamingTypeOptions
-      }
-    }
-  },
-  {
-    key: "sponsor",
-    label: "Sponsor",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "company",
-      props: {
-        multiple: true
-      }
-    },
-    customParser: (f) => [
-      `sponsor==${f.value.map((s) => escapeFilterValue(s.raw.name)).join("||")}`
-    ]
-  },
-  {
-    key: "all_companies",
-    label: "All Companies",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "company",
-      props: {
-        queryFunction: queryAllCompanies,
-        multiple: true
-      }
-    },
-    customParser: (f) => {
-      const companies = f.value
-        .map((c) => escapeFilterValue(c.raw.name))
-        .join("||");
-      return [
-        `speaker_company==${companies},created_by_company==${companies},sponsor==${companies}`
-      ];
-    }
-  },
-  {
-    key: "submission_status",
-    label: "Submitter Company",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: submissionStatusOptions
-      }
-    }
-  },
-  {
-    key: "media_upload_with_type",
-    label: "Media Upload Type",
-    operators: [OPERATORS.HAS, OPERATORS.HAS_NOT],
-    values: {
-      type: "select",
-      props: {
-        options: mediaUploadTypes.map((type) => ({
-          value: type.id,
-          label: type.name
-        })),
-        multiple: true
-      }
-    },
-    customParser: (f) => {
-      const filter = [];
-
-      if (f.operator === OPERATORS.HAS.value) {
-        const value = Array.isArray(f.value) ? f.value.join("||") : f.value;
-        filter.push(`has_media_upload_with_type==${value}`);
-      } else {
-        const value = Array.isArray(f.value) ? f.value.join("&&") : f.value;
-        filter.push(`has_not_media_upload_with_type==${value}`);
-      }
-
-      return filter;
-    }
-  },
-  {
-    key: "review_status",
-    label: "Review Status",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: reviewStatusOptions,
-        multiple: true
-      }
-    }
-  },
-  {
-    key: "submission_source",
-    label: "Submitter Status",
-    operators: [OPERATORS.IS],
-    values: {
-      type: "select",
-      props: {
-        options: submissionSourceOptions
-      }
-    }
-  }
-];
 
 const SummitEventListPage = ({
   events,
@@ -1132,7 +110,7 @@ const SummitEventListPage = ({
       orderDir: od
     } = mergedParams;
 
-    getEvents(t, p, pp, o, od, parsedFilter, extraColumns);
+    getEvents(t, p, pp, o, od, parsedFilter, selectedColumns);
   };
 
   useEffect(() => {
@@ -1181,25 +159,7 @@ const SummitEventListPage = ({
   };
 
   const handleSort = (index, key, dir) => {
-    let translatedKey = key;
-    switch (key) {
-      case "name":
-        translatedKey = "last_name";
-        break;
-      case "submitter_company":
-        translatedKey = "created_by_company";
-        break;
-      case "progress_flags":
-        translatedKey = "actions";
-        break;
-      case "track_name":
-        translatedKey = "track";
-        break;
-      default:
-        break;
-    }
-
-    _getEvents({ order: translatedKey, orderDir: dir });
+    _getEvents({ order: toApiSortKey(key), orderDir: dir });
   };
 
   const handleSearch = (newTerm) => {
@@ -1244,35 +204,58 @@ const SummitEventListPage = ({
     saveFilterCriteria(filterToSave);
   };
 
+  const optionalColumns = getOptionalColumns(
+    currentSummit.selection_plans,
+    currentSummit.tracks,
+    currentSummit.event_types,
+    currentSummit.id
+  );
+
+  const columnDDLOptions = [
+    ...optionalColumns.map((oc) => ({ value: oc.columnKey, label: oc.label })),
+    {
+      value: "all_companies",
+      label: T.translate("event_list.all_companies")
+    },
+    {
+      value: "all",
+      label: T.translate("general.all")
+    }
+  ];
+
   const handleColumnsChange = (ev) => {
-    const { value } = ev.target;
-    let newColumns = value;
+    const { value: newColumns } = ev.target;
+
+    if (newColumns.includes("all")) {
+      setSelectedColumns(
+        columnDDLOptions.map((opt) => opt.value).filter((v) => v !== "all")
+      );
+      return;
+    }
+
     const allCompanies = ["submitter_company", "speaker_company", "sponsor"];
-    const mediaUploadsSelected = newColumns.includes("media_uploads");
-
-    newColumns = newColumns.filter((col) => col !== "media_uploads_display");
-    if (mediaUploadsSelected) newColumns.push("media_uploads_display");
-
+    const hadAllCompanies = selectedColumns.includes("all_companies");
     const hasAllCompanies = newColumns.includes("all_companies");
-    const selectedCompanies = selectedColumns.filter((c) =>
-      allCompanies.includes(c)
-    ).length;
-    const newCompanies = newColumns.filter((c) =>
-      allCompanies.includes(c)
-    ).length;
 
-    if (selectedColumns.includes("all_companies") && !hasAllCompanies) {
-      newColumns = newColumns.filter((e) => !allCompanies.includes(e));
-    } else if (hasAllCompanies) {
-      if (newCompanies === 0) {
-        newColumns = [...selectedColumns, ...allCompanies, "all_companies"];
-      } else if (newCompanies === selectedCompanies) {
-        newColumns = [
-          ...new Set([...newColumns, ...allCompanies, "all_companies"])
-        ];
-      } else if (newCompanies < selectedCompanies) {
-        newColumns = newColumns.filter((c) => c !== "all_companies");
-      }
+    if (hadAllCompanies && !hasAllCompanies) {
+      setSelectedColumns(newColumns.filter((c) => !allCompanies.includes(c)));
+      return;
+    }
+
+    if (hasAllCompanies) {
+      const selectedCompanies = allCompanies.filter((c) =>
+        selectedColumns.includes(c)
+      ).length;
+      const newCompanies = allCompanies.filter((c) =>
+        newColumns.includes(c)
+      ).length;
+
+      setSelectedColumns(
+        newCompanies < selectedCompanies
+          ? newColumns.filter((c) => c !== "all_companies")
+          : [...new Set([...newColumns, ...allCompanies])]
+      );
+      return;
     }
 
     setSelectedColumns(newColumns);
@@ -1288,30 +271,13 @@ const SummitEventListPage = ({
     });
   };
 
-  const translateSortKey = (key) => {
-    switch (key) {
-      case "last_name":
-        return "name";
-      case "created_by_company":
-        return "submitter_company";
-      case "actions":
-        return "progress_flags";
-      case "track":
-        return "track_name";
-      default:
-        break;
-    }
-
-    return key;
-  };
-
   const eventTypeOptions = buildNameIdDDL(currentSummit.event_types);
 
-  let columns = [
-    { columnKey: "id", value: T.translate("general.id"), sortable: true },
+  const fixedColumns = [
+    { columnKey: "id", label: T.translate("general.id"), sortable: true },
     {
       columnKey: "type",
-      value: T.translate("event_list.type"),
+      label: T.translate("event_list.type"),
       sortable: true,
       // eslint-disable-next-line react/no-unstable-nested-components
       editableField: (extraProps) => (
@@ -1337,14 +303,14 @@ const SummitEventListPage = ({
     },
     {
       columnKey: "title",
-      value: T.translate("event_list.title"),
+      label: T.translate("event_list.title"),
       sortable: true,
       editableField: true,
       placeholder: T.translate("bulk_actions_page.placeholders.event_title")
     },
     {
       columnKey: "selection_status",
-      value: T.translate("event_list.selection_status"),
+      label: T.translate("event_list.selection_status"),
       sortable: true,
       render: (status, row) =>
         status === "unaccepted" && row.is_published === true
@@ -1354,7 +320,7 @@ const SummitEventListPage = ({
   ];
 
   const tableOptions = {
-    sortCol: translateSortKey(order),
+    sortCol: toUiSortKey(order),
     sortDir: orderDir,
     className: "summit-event-list-table",
     actions: {
@@ -1363,117 +329,11 @@ const SummitEventListPage = ({
     }
   };
 
-  const columnOptions = [
-    {
-      value: "event_type_capacity",
-      label: T.translate("event_list.event_type_capacity")
-    },
-    { value: "speaker_names", label: T.translate("event_list.speakers") },
-    {
-      value: "all_companies",
-      label: T.translate("event_list.all_companies")
-    },
-    {
-      value: "created_by_fullname",
-      label: T.translate("event_list.created_by")
-    },
-    { value: "duration", label: T.translate("event_list.duration") },
-    { value: "end_date", label: T.translate("event_list.end_date") },
-    { value: "published_date", label: T.translate("event_list.published") },
-    {
-      value: "speaker_company",
-      label: T.translate("event_list.speaker_company")
-    },
-    {
-      value: "speakers_count",
-      label: T.translate("event_list.speakers_count")
-    },
-    { value: "sponsor", label: T.translate("event_list.sponsor") },
-    {
-      value: "selection_plan",
-      label: T.translate("event_list.selection_plan")
-    },
-    { value: "location", label: T.translate("event_list.location") },
-    { value: "level", label: T.translate("event_list.level") },
-    { value: "tags", label: T.translate("event_list.tags") },
-    {
-      value: "streaming_url",
-      label: T.translate("event_list.streaming_url")
-    },
-    { value: "meeting_url", label: T.translate("event_list.meeting_url") },
-    {
-      value: "etherpad_link",
-      label: T.translate("event_list.etherpad_link")
-    },
-    {
-      value: "streaming_type",
-      label: T.translate("event_list.streaming_type")
-    },
-    { value: "start_date", label: T.translate("event_list.start_date") },
-    {
-      value: "submitter_company",
-      label: T.translate("event_list.submitter_company")
-    },
-    { value: "track_name", label: T.translate("event_list.track") },
-    { value: "status", label: T.translate("event_list.submission_status") },
-    {
-      value: "submission_source",
-      label: T.translate("event_list.submission_source")
-    },
-    {
-      value: "progress_flags",
-      label: T.translate("event_list.progress_flags")
-    },
-    {
-      value: "media_uploads",
-      label: T.translate("event_list.media_uploads")
-    },
-    {
-      value: "review_status",
-      label: T.translate("event_list.review_status")
-    },
-    { value: "created", label: T.translate("event_list.created") },
-    { value: "modified", label: T.translate("event_list.modified") },
-    {
-      value: "allow_feedback",
-      label: T.translate("event_list.allow_feedback")
-    },
-    {
-      value: "to_record",
-      label: T.translate("event_list.to_record")
-    }
-  ];
+  const selectedOptionalColumns = optionalColumns.filter((c) =>
+    selectedColumns.includes(c.columnKey)
+  );
 
-  const showColumns = fieldNames(
-    currentSummit.selection_plans,
-    currentSummit.tracks,
-    currentSummit.event_types,
-    currentSummit.id
-  )
-    .filter(
-      (f) =>
-        selectedColumns.includes(f.columnKey) &&
-        !defaultColumns.includes(f.columnKey)
-    )
-    .map((f2) => {
-      let c = {
-        columnKey: f2.columnKey,
-        value: T.translate(`event_list.${f2.value}`),
-        sortable: f2.sortable,
-        editable: !!editableColumns.includes(f2.editable),
-        customStyle: f2.customStyle
-      };
-      // optional fields
-      if (f2?.title) c = { ...c, title: f2.title };
-
-      if (f2?.render) c = { ...c, render: f2.render };
-
-      if (f2?.editableField) c = { ...c, editableField: f2.editableField };
-
-      return c;
-    });
-
-  columns = [...columns, ...showColumns];
+  const tableColumns = [...fixedColumns, ...selectedOptionalColumns];
 
   if (!currentSummit.id) return <div />;
 
@@ -1560,7 +420,7 @@ const SummitEventListPage = ({
             placeholder={T.translate("event_list.placeholders.select_fields")}
             value={selectedColumns}
             onChange={handleColumnsChange}
-            options={handleDDLSortByLabel(columnOptions)}
+            options={handleDDLSortByLabel(columnDDLOptions)}
             isClearable
             isMulti
           />
@@ -1575,9 +435,9 @@ const SummitEventListPage = ({
             <BulkEditTable
               options={tableOptions}
               data={tableData}
-              columns={columns}
+              columns={tableColumns}
               onSort={handleSort}
-              onUpdate={(rows) => bulkUpdateEvents(currentSummit.id, rows)}
+              onUpdate={bulkUpdateEvents}
             />
           </div>
           <Pagination
